requirements.txt
# Python dependencies for AI Translator & Critic

# Web framework
Flask>=2.3.0

# HTTP requests
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-mock>=3.11.0
